relocated api to parent component and added clean up
