+++ src/screens/App/screens/Seller/screens/Orders/screens/Master/components/OrdersTable/OrderTable.jsx
@@ -1,68 +1,12 @@
-import React, { useEffect } from "react";
-import { useHistory } from "react-router-dom";
-import { batch, useDispatch, useSelector } from "react-redux";
+import React from "react";
+import { useSelector } from "react-redux";
 import Loading from "../../../../../../../../../../shared/Loading/Loading";
 import OrderHeading from "./OrderHeading";
 import OrderRow from "./OrderRow";
-import axios from "../../../../../../../../../../shared/caller";
-import {
-  setMessage,
-  setSeverity,
-} from "../../../../../../../../../../redux/Alert/AlertAction";
-import {
-  loadOrders,
-  updatePageLoading,
-} from "../../../../../../../../../../redux/Seller/OrdersMaster/OrdersMasterAction";
 
 const OrderTable = () => {
-  const history = useHistory();
-
-  // redux
-  const dispatch = useDispatch();
-
   // redux orders master reducer & state
   const loading = useSelector((state) => state.ORDERS_MASTER.pageLoading);
-  const status = useSelector((state) => state.ORDERS_MASTER.status);
-
-  // triggered every time status query changes
-  useEffect(() => {
-    async function getOrders() {
-      await axios
-        .get(`/api/seller/orders/master/${status}`)
-        .then((res) => {
-          if (res.status === 200) {
-            const orders = res.data.orders;
-
-            batch(() => {
-              dispatch(updatePageLoading(false));
-              dispatch(loadOrders(orders));
-            });
-          }
-        })
-        .catch((err) => {
-          if (!err.response)
-            batch(() => {
-              dispatch(setSeverity("error"));
-              dispatch(
-                setMessage(
-                  "Something went wrong in retrieving your orders. Refresh your browser or try again later."
-                )
-              );
-              dispatch(updatePageLoading(false));
-            });
-          else if (err.response.status === 403) history.push("/forbidden");
-          else if (err.response.status === 401) history.push("unathorized");
-          else
-            batch(() => {
-              dispatch(setSeverity("error"));
-              dispatch(setMessage(err.response.data.error));
-              dispatch(updatePageLoading(false));
-            });
-        });
-    }
-
-    getOrders();
-  }, [status]);
 
   return (
     <table className="w-full table-fixed">
